requirements.txt
requests>=2.31
python-dotenv>=1.0
beautifulsoup4>=4.12
lxml>=5.0
pandas>=2.2
psycopg2-binary>=2.9
streamlit>=1.35
streamlit-autorefresh>=1.0
plotly>=5.22
pytz>=2024.1
